scripts/requirements.txt
Pillow
lxml
requests
qrcode[pil]
argparse
adafruit-ampy

